[write] Move layout builder types to layout::builders
We will be adding a bunch more builders shortly.
